feat(ci): actions to connect to codeartifact feed
A workflow approach was attempted but it proved difficult to adequately
mask the token from workflow run logs.

Various masking strategies were tried here to prevent the JWT from
appearing in the log.

  - adding a mask to the value output from the aws/codeartifact-connect
    prevents it being used as an output
  - adding a mask asap in the calling workflow does work, without
    preventing it being used as an output, but it's still going to be
    displayed in the log when referencing the output.

The JWT is scoped to the GitHub account, but further hardening may be
required. A composite action approach keeps it out of the logs and can
further benefit from bundles scripts.

Additionally, the workflow was very quick so instead of a billable, but
short running job the action means it can be executed inline.